+++ xcube/cli/serve.py
@@ -22,7 +22,7 @@ from typing import List
 
 import click
 
-from xcube.webapi.defaults import DEFAULT_PORT, DEFAULT_NAME, DEFAULT_ADDRESS, DEFAULT_UPDATE_PERIOD, \
+from xcube.webapi.defaults import DEFAULT_PORT, DEFAULT_ADDRESS, DEFAULT_UPDATE_PERIOD, \
     DEFAULT_TILE_CACHE_SIZE, DEFAULT_TILE_COMP_MODE
 
 __author__ = "Norman Fomferra (Brockmann Consult GmbH)"
@@ -32,18 +32,21 @@ VIEWER_ENV_VAR = 'XCUBE_VIEWER_PATH'
 
 @click.command(name='serve')
 @click.argument('cubes', metavar='CUBE...', nargs=-1)
-@click.option('--name', '-n', metavar='NAME', default=DEFAULT_NAME,
-              help=f'Service name. Defaults to {DEFAULT_NAME!r}.')
 @click.option('--address', '-a', metavar='ADDRESS', default=DEFAULT_ADDRESS,
               help=f'Service address. Defaults to {DEFAULT_ADDRESS!r}.')
 @click.option('--port', '-p', metavar='PORT', default=DEFAULT_PORT, type=int,
               help=f'Port number where the service will listen on. Defaults to {DEFAULT_PORT}.')
+@click.option('--prefix', metavar='PREFIX',
+              help='Service URL prefix. May contain template patterns such as "${version}" or "${name}". '
+                   'For example "${name}/api/${version}".')
+@click.option('--name', '-n', metavar='NAME', hidden=True,
+              help='Service name. Deprecated, use prefix option instead.')
 @click.option('--update', '-u', metavar='PERIOD', type=float,
               default=DEFAULT_UPDATE_PERIOD,
               help='Service will update after given seconds of inactivity. Zero or a negative value will '
                    'disable update checks. '
-                   f'Defaults to {DEFAULT_UPDATE_PERIOD!r}.')
-@click.option('--styles', '-s', metavar='STYLES', default=None,
+              f'Defaults to {DEFAULT_UPDATE_PERIOD!r}.')
+@click.option('--styles', '-S', metavar='STYLES', default=None,
               help='Color mapping styles for variables. '
                    'Used only, if one or more CUBE arguments are provided and CONFIG is not given. '
                    'Comma-separated list with elements of the form '
@@ -53,25 +56,26 @@ VIEWER_ENV_VAR = 'XCUBE_VIEWER_PATH'
                    'Cannot be used if CUBES are provided.')
 @click.option('--tilecache', metavar='SIZE', default=DEFAULT_TILE_CACHE_SIZE,
               help=f'In-memory tile cache size in bytes. '
-                   f'Unit suffixes {"K"!r}, {"M"!r}, {"G"!r} may be used. '
-                   f'Defaults to {DEFAULT_TILE_CACHE_SIZE!r}. '
-                   f'The special value {"OFF"!r} disables tile caching.')
+              f'Unit suffixes {"K"!r}, {"M"!r}, {"G"!r} may be used. '
+              f'Defaults to {DEFAULT_TILE_CACHE_SIZE!r}. '
+              f'The special value {"OFF"!r} disables tile caching.')
 @click.option('--tilemode', metavar='MODE', default=None, type=int,
               help='Tile computation mode. '
                    'This is an internal option used to switch between different tile computation implementations. '
-                   f'Defaults to {DEFAULT_TILE_COMP_MODE!r}.')
+              f'Defaults to {DEFAULT_TILE_COMP_MODE!r}.')
 @click.option('--show', '-s', is_flag=True,
               help=f"Run viewer app. Requires setting the environment variable {VIEWER_ENV_VAR} "
-                   f"to a valid xcube-viewer deployment or build directory. "
-                   f"Refer to https://github.com/dcs4cop/xcube-viewer for more information.")
+              f"to a valid xcube-viewer deployment or build directory. "
+              f"Refer to https://github.com/dcs4cop/xcube-viewer for more information.")
 @click.option('--verbose', '-v', is_flag=True,
               help="Delegate logging to the console (stderr).")
 @click.option('--traceperf', is_flag=True,
               help="Print performance diagnostics (stdout).")
 def serve(cubes: List[str],
-          name: str,
           address: str,
           port: int,
+          prefix: str,
+          name: str,
           update: float,
           styles: str,
           config: str,
@@ -89,6 +93,8 @@ def serve(cubes: List[str],
 
     from xcube.util.cliutil import parse_cli_kwargs
 
+    prefix = prefix or name
+
     if config and cubes:
         raise click.ClickException("CONFIG and CUBES cannot be used at the same time.")
     if styles:
@@ -102,8 +108,8 @@ def serve(cubes: List[str],
 
     from xcube.webapi.app import new_application
     from xcube.webapi.service import Service
-    service = Service(new_application(name),
-                      name=name,
+    service = Service(new_application(prefix),
+                      prefix=prefix,
                       port=port,
                       address=address,
                       cube_paths=cubes,