+++ xcube/webapi/handlers.py
@@ -37,20 +37,6 @@ from ..util.timecoord import timestamp_to_iso_string
 
 __author__ = "Norman Fomferra (Brockmann Consult GmbH)"
 
-_WMTS_KVP_KEYS = [
-    'Service',
-    'Request',
-    'Version',
-    'Format',
-    'Style',
-    'Layer',
-    'TileMatrixSet',
-    'TileMatrix',
-    'TileRow',
-    'TileCol'
-]
-
-_WMTS_KVP_LOWER_KEYS = [k.lower() for k in _WMTS_KVP_KEYS]
 _WMTS_VERSION = "1.0.0"
 _WMTS_TILE_FORMAT = "image/png"
 
@@ -59,8 +45,8 @@ _WMTS_TILE_FORMAT = "image/png"
 class WMTSKvpHandler(ServiceRequestHandler):
 
     async def get(self):
-        # According to WMTS 1.0 spec, all WMTS-specific keys must be case insensitive.
-        self._convert_wmts_keys_to_lower_case()
+        # According to WMTS 1.0 spec, query parameters must be case-insensitive.
+        self.set_caseless_query_arguments()
 
         service = self.params.get_query_argument('service')
         if service != "WMTS":
@@ -107,17 +93,6 @@ class WMTSKvpHandler(ServiceRequestHandler):
         else:
             raise ServiceBadRequestError(f'Invalid request type "{request}"')
 
-    def _convert_wmts_keys_to_lower_case(self):
-        query_arguments = dict(self.request.query_arguments)
-        query_keys = {k.lower(): k for k in query_arguments.keys()}
-        for lower_key in _WMTS_KVP_LOWER_KEYS:
-            if lower_key in query_keys:
-                query_key = query_keys[lower_key]
-                value = query_arguments[query_key]
-                del query_arguments[query_key]
-                query_arguments[lower_key] = value
-        self.request.query_arguments = query_arguments
-
 
 # noinspection PyAbstractClass
 class GetWMTSCapabilitiesXmlHandler(ServiceRequestHandler):
@@ -158,6 +133,21 @@ class GetDatasetCoordsHandler(ServiceRequestHandler):
         response = get_dataset_coordinates(self.service_context, ds_id, dim_name)
         self.set_header('Content-Type', 'application/json')
         self.write(json.dumps(response, indent=2))
+
+
+# noinspection PyAbstractClass,PyBroadException
+class GetWMTSTileHandler(ServiceRequestHandler):
+
+    async def get(self, ds_id: str, var_name: str, z: str, y: str, x: str):
+        self.set_caseless_query_arguments()
+        tile = await IOLoop.current().run_in_executor(None,
+                                                      get_dataset_tile,
+                                                      self.service_context,
+                                                      ds_id, var_name,
+                                                      x, y, z,
+                                                      self.params)
+        self.set_header('Content-Type', 'image/png')
+        self.finish(tile)
 
 
 # noinspection PyAbstractClass,PyBroadException