+++ CHANGES.md
@@ -1,5 +1,25 @@
 ## Changes in 0.2.0 (in dev)
 
+### Enhancements
+
+* Reworked `xcube gen` to be more user-friendly and more consistent with other tools. 
+  The changes are
+  - Removed `--dir` and `--name` options and replaced it by single `--output` option, 
+    whose default value is `out.zarr`. (#45)
+  - The `--format` option no longer has a default value. If not given, 
+    format is guessed from `--output` option.
+  - Renamed following parameters in the configuration file:
+    + `input_files` into `input_paths`, also because paths may point into object storage 
+      locations (buckets);  
+    + `output_file` into `output_path`, to be consistent with `input_paths`.  
+* Added new CLI command `xcube prune`. The tool deletes all block files associated with empty (NaN-
+  only) chunks in given INPUT cube, which must have ZARR format. This can drastically reduce files 
+  in sparse cubes and improve cube reading performance. (#92)
+* `xcube serve` has a new `prefix` option which is a path appended to the server's host.
+  The `prefix` option replaces the `name` option which is now deprecated but kept 
+  for backward compatibility. (#79)
+* Added new CLI command `xcube resample` that is used to generate temporarily up- or downsampled
+  data cubes from other data cubes.
 * `xcube serve` can now be run with data cube paths and styling information given via the CLI rather 
   than a configuration file. For example `xcube serve --styles conc_chl=(0,20,"viridis") /path/to/my/chl-cube.zarr`.
   This allows for quick inspection of newly generated cubes via `xcube gen`.
@@ -20,13 +40,21 @@
 * Added `xcube gen` option `--sort` when input data list should be sorted (#33)    
 * Added `xcube vars2dim` command to make variables a cube dimension (#31)
 * Added `xcube serve` option `--traceperf` that allows switching on performance diagnostics.
-* Fixed error in plugins when importing `xcube.api.gen` (#62)
-* Fixed import of plugins only when executing `xcube.cli` (#66)
 * Included possibility to read the input file paths from a text file. (#47)
 * Restructured and clarified code base (#27)
 * Moved to Python 3.7 (#25)
 * Excluding all input processors except for the default one. They are now plugins and have own repositories within the 
 xcube's organisation. (#49)
+
+### Fixes
+
+* Fixed problem where CTRL+C didn't function anymore with `xcube serve`. (#87)
+* Fixed error `indexes along dimension 'y' are not equal` occurred when using 
+  `xcube gen` with processed variables that used flag values (#86)
+* Fixed `xcube serve` WMTS KVP API to allow for case-insensitive query parameters. (#77)
+* Fixed error in plugins when importing `xcube.api.gen` (#62)
+* Fixed import of plugins only when executing `xcube.cli` (#66)
+
 
 ## Changes in 0.1.0